// frontend/src/types/context/auth.ts
export interface ICompanyInfo {
  company_name: string;
  company_description: string;
  company_logo: string;
}
// ...
export interface IUpdateContactInfo {
  first_name: string;
  last_name: string;
  location?: IUpdateLocation;
  company?: ICompanyInfo;
}
export interface IUpdateLocation {
  address: string;
  lat: string;
  lngs: string;
  time_zone: string;
  country: string;
  street_address: string;
  city: string;
  postal_code: string;
}